+++ share/auto_configure.sh
@@ -18,7 +18,7 @@ catchall() {
 	if [ $rc -gt 0 ] ; then
 		fatal "Failure. See ${LOGFILE} for details."
 	else
-		rm -f ${LOGFILE}
+		rm -f "${LOGFILE}"
 	fi
 	exec 3>&-
 	trap - INT EXIT TERM
@@ -26,7 +26,7 @@ catchall() {
 
 
 fatal() {
-	echo -e "\e[1;31m$@\e[0m" | tee -a /dev/fd/3 >&2
+	echo -e "\\e[1;31m$*\\e[0m" | tee -a /dev/fd/3 >&2
 	exit 1
 }
 
@@ -40,8 +40,8 @@ setup_logging() {
 	if [ -n "${DEBUG-}" ] ; then
 		exec 3>/dev/null
 	else
-		exec 3>&2 2>${LOGFILE} 1>&2
-		chmod 0600 ${LOGFILE}
+		exec 3>&2 2>"$LOGFILE" 1>&2
+		chmod 0600 "$LOGFILE"
 		trap 'catchall' INT EXIT TERM
 	fi
 
@@ -53,11 +53,10 @@ setup_logging() {
 setup_pq() {
 	# Ensure used libpq vars are defined for configuration template.
 
-	export PGHOST=${PGHOST-$(query_pgsettings unix_socket_directories)}
-	PGHOST=${PGHOST%%,*}
+	export PGHOST=${PGHOST-/var/run/postgresql}
 	export PGPORT=${PGPORT-5432}
 	export PGUSER=${PGUSER-postgres}
-	if [ -d ${PGHOST-/tmp} ] ; then
+	if [ -d "$PGHOST" ] ; then
 		sudo="sudo -Eu ${PGUSER}"
 	else
 		sudo=
@@ -80,7 +79,7 @@ setup_ssl() {
 		fatal "Failed to find PKI directory."
 	fi
 
-	if [ -f $pki/certs/ssl-cert-snakeoil.pem -a -f $pki/private/ssl-cert-snakeoil.key ] ; then
+	if [ -f $pki/certs/ssl-cert-snakeoil.pem ] && [ -f $pki/private/ssl-cert-snakeoil.key ] ; then
 		log "Using snake-oil SSL certificate."
 		sslcert=$pki/certs/ssl-cert-snakeoil.pem
 		sslkey=$pki/private/ssl-cert-snakeoil.key
@@ -95,14 +94,13 @@ setup_ssl() {
 		     chmod 644 $sslkey
 	     fi
 	fi
-	echo $sslcert $sslkey
+	readlink -e $sslcert $sslkey
 }
 
 
 generate_configuration() {
 	local sslcert=$1; shift
 	local sslkey=$1; shift
-	local cookiesecret="$(pwgen 128)"
 
 	cat <<-EOF
 	# Configuration initiated by $0 on $(date)
@@ -113,7 +111,7 @@ generate_configuration() {
 	[temboard]
 	ssl_cert_file = $sslcert
 	ssl_key_file = $sslkey
-	cookie_secret = ${cookiesecret}
+	cookie_secret = $(pwgen 128)
 	home = ${VARDIR}
 
 	[repository]
@@ -144,7 +142,7 @@ pwgen() {
 
 #       M A I N
 
-cd $(readlink -m ${BASH_SOURCE[0]}/..)
+cd "$(readlink -m "${BASH_SOURCE[0]}/..")"
 
 setup_logging
 setup_pq
@@ -153,8 +151,8 @@ export TEMBOARD_PASSWORD=${TEMBOARD_PASSWORD-$(pwgen)}
 if ! getent passwd temboard ; then
 	log "Creating system user temBoard."
 	useradd \
-		--system --user-group --shell /sbin/nologin \
-		--home-dir ${VARDIR} \
+		--system --user-group --shell "$SHELL" \
+		--home-dir "$VARDIR" \
 		--comment "temBoard Web UI" temboard &>/dev/null
 fi
 
@@ -164,18 +162,14 @@ fi
 
 
 log "Configuring temboard in ${ETCDIR}."
-sslfiles=($(set -eu; setup_ssl))
-install -o temboard -g temboard -m 0750 -d ${ETCDIR} ${LOGDIR} ${VARDIR}
-install -o temboard -g temboard -m 0640 /dev/null ${ETCDIR}/temboard.conf
-generate_configuration "${sslfiles[@]}" > ${ETCDIR}/temboard.conf
+mapfile sslfiles < <(set -eu; setup_ssl)
+install -o temboard -g temboard -m 0750 -d "$ETCDIR" "$LOGDIR" "$VARDIR"
+install -o temboard -g temboard -m 0640 /dev/null "$ETCDIR/temboard.conf"
+generate_configuration "${sslfiles[@]}" > "$ETCDIR/temboard.conf"
 
 log "Creating Postgres user, database and schema."
-if [ -d ${PGHOST-/tmp} ] ; then
-	# If local, sudo to PGUSER.
-	sudo -Eu ${PGUSER} ./create_repository.sh
-else
-	./create_repository.sh
-fi
+./create_repository.sh
+
 dsn="postgres://temboard:${TEMBOARD_PASSWORD}@/temboard"
 if ! sudo -Eu temboard psql -Atc "SELECT 'CONNECTED';" "$dsn" | grep -q 'CONNECTED' ; then
 	fatal "Can't configure access to Postgres database."