+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## [7.10] - Unreleased
+
+- Fix database ownership on creation.
+- Ship a /usr/share/temboard/reassign.sql script to fix ownership.
+- Define a shell to temboard UNIX user.
+- Review defaults groups.
+
 
 ## [7.9] - 2022-01-03